power: Save battery by reducing unneccesary network calls
When an app makes frequent network calls (which isn't neccessary), it drains more battery. The idea here is to
reduce such network calls and usage by refreshing data only from the MainActivity. Once the data refresh is
done, it gets stored offlinefor further use.

Whenever user clicks on the changelog, the app starts to verify whether an offline data is available to load.
If no data persists in the storage, then the user needs to refresh the data from the MainActivity(otherwise, the
user will end up seeing blank data) thus saving bandwidth and power of the device. There should be great improvement
in loading the data when device is offline because the datagets loaded from the device.

Signed-off-by: AxelBlaz3 <karthikgaddam4@gmail.com>