Perform a check when Nano is installed without properties
There's a case where Nano is installed by the user but the properties that the app uses to retrieve kernel info are not yet present in build.prop.

It's most likely because the user's device isn't running latest Nano. To fix this, Updater now shows the latest build available as an update to the user. When user flashes the latest build, Updater will detect the properties as they're appended to build.prop

Signed-off-by: AxelBlaz3 <karthikgaddam4@gmail.com>