Run startup object initialization in background threads
There's a problem when the onCreate method is overloaded with initializations. Views are displayed only when the initializations are done. Hence it results in a blank screen for few seconds (till the initializations are done).

Such object initializations, when executed in a background thread will improve the startup speed of the app thereby resulting in faster view displays. The data related to such views gets loaded when the background thread is done with it's execution. This is what so called 'lazy loading'.

Signed-off-by: AxelBlaz3 <karthikgaddam4@gmail.com>