Improvments to startup speed and bug fixes
Replace <include> layout with ViewStub. As ViewStub is a zero dimensional View, it doesn't need GPU to draw the views. This inturn leads to nearly 30-40% improvement in the app's startup time (at least in this case) because two layouts will not be drawn at the app startup.

There's a bug in the FlashActivity. Clicking on the FAB and then selecting the zip will lead to a crash. This occurs when the Uri data from the intent isn't formatted according to the logic. This fix includes a new way to retrieve the path of selected file which is more reliable.

Signed-off-by: AxelBlaz3 <karthikgaddam4@gmail.com>