Verify if the package is AnyKernel zip
A user might select a zip which is not a kernel. For instance, selecting a ROM zip might be risky as the flasher tries to flash the ROM within the app which might make device unusable. To avoid such issues, Updater now verifies if the selected package is a valid AnyKernel zip. if it isn't, then the Updater simply shows a snackbar message "Choose a valid AnyKernel zip".

Signed-off-by: AxelBlaz3 <karthikgaddam4@gmail.com>