Revert "Use an efficient approach to handle zip selection"
This reverts commit a68b28b4a7e00b3c7c298560645a3f21f24f40a0. This commit focuses on supporting Updater with android 10.
Reason for revert:
Android 10 brings new privacy changes for devices. One of them is so what called "Scoped Storage" which makes sure that an app can access only it's specific directories (Android/data/myapp or cache dir...) and hence loses it's capability for accessing internal storage. The only way to access files on device is to use Storage Access Framework (SAF) as per documentation (or at least as of now).

Signed-off-by: AxelBlaz3 <karthikgaddam4@gmail.com>

# Conflicts:
#	app/src/main/java/com/codebot/axel/kernel/updater/FlashActivity.kt
#	app/src/main/java/com/codebot/axel/kernel/updater/util/Constants.kt
#	app/src/main/java/com/codebot/axel/kernel/updater/util/FileChooser.kt
#	app/src/main/java/com/codebot/axel/kernel/updater/util/FlashKernel.kt
#	app/src/main/res/drawable-mdpi/chooser_background.xml
#	app/src/main/res/values/styles.xml